extract print_diff_to_console to utils
