+++ src/snowflake/cli/_plugins/nativeapp/manager.py
@@ -62,11 +62,11 @@ from snowflake.cli._plugins.stage.diff import (
     StagePath,
     compute_stage_diff,
     preserve_from_diff,
-    print_diff_to_console,
     sync_local_diff_with_stage,
     to_stage_path,
 )
 from snowflake.cli._plugins.stage.manager import StageManager
+from snowflake.cli._plugins.stage.utils import print_diff_to_console
 from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.console import cli_console as cc
 from snowflake.cli.api.errno import (
